[feat] 예매 페이지 이동 navigation 연결 #63
